# backend/docs/swagger.yaml
components:
  schemas:
    data:
      properties:
        data:
          $ref: '#/components/schemas/models.UserResponse'
      type: object
...
    models.ErrorResponse:
      description: Generic API error response
      properties:
        code:
          type: integer
        errors:
          items:
            $ref: '#/components/schemas/models.Error'
          type: array
          uniqueItems: false
        message:
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          type: string
      type: object
...
    models.Meta:
      description: Meta object
      properties:
        timestamp:
          example: "2025-02-12T18:31:40Z"
          type: string
      type: object
...
    models.ShopCreateParams:
      properties:
        currency_code:
          enum:
          - USD
          - NGN
          type: string
        status:
          enum:
          - PUBLISHED
          - DRAFT
          type: string
        subdomain:
          maxLength: 255
          minLength: 3
          type: string
        template:
          type: string
        title:
          maxLength: 255
          minLength: 3
          type: string
      required:
      - currency_code
      - status
      - subdomain
      - template
      - title
      type: object
...
    models.SuccessResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      description: Generic API response
      properties:
        code:
          example: 200
          type: integer
        data: {}
        message:
          example: Object updated successfully
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          example: success
          type: string
      type: object
...
  securitySchemes:
    OAuth2AccessCode:
      flows:
        authorizationCode:
          authorizationUrl: http://127.0.0.1:8080/oauth2/auth
          tokenUrl: http://127.0.0.1:8080/oauth2/token
      in: header
      type: oauth2
...
info:
  description: This is the Naytife API documentation
  title: Naytife API Docs
  version: "1.0"
openapi: 3.1.0
paths:
...
  /shops:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Shops fetched successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Fetch all shops
      tags:
      - shops
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.ShopCreateParams'
        description: Shop object that needs to be created
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Shop created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Create a shop
      tags:
      - shops
  /shops/{shop_id}:
    delete:
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.SuccessResponse'
          description: Shop deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Delete a shop
      tags:
      - shops
    get:
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Shop fetched successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Fetch a shop
      tags:
      - shops
    put:
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.ShopUpdateParams'
        description: Shop update parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Shop updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Update a shop
      tags:
      - shops
...
  /shops/{shop_id}/deploy:
    post:
      description: Deploy a store using the store deployer service
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.StoreDeploymentRequest'
        description: Deployment parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.DeploymentResponse'
          description: Store deployed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Deploy a store
      tags:
      - deployment
...
  /shops/{shop_id}/images:
    put:
      description: Update shop images (logo, favicon, banner, etc)
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.ShopImagesUpdateParams'
        description: Shop images object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Shop images updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Bad request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal server error
      security:
      - OAuth2AccessCode: []
      summary: Update shop images
      tags:
      - shops
...
servers:
- url: http://127.0.0.1:8080/v1